+++ django_spire/core/management/commands/spire_startapp_pkg/template/app/tests/test_intelligence/test_bots.py
@@ -1,6 +1,0 @@
-from django.test import TestCase
-
-
-class SpireChildAppBotTestCase(TestCase):
-    def setUp(self):
-        super().setUp()