+++ django_spire/core/management/commands/spire_startapp_pkg/template/app/tests/test_services/test_transformation_service.py
@@ -1,6 +1,0 @@
-from django.test import TestCase
-
-
-class SpireChildAppTransformationServiceTestCase(TestCase):
-    def setUp(self):
-        super().setUp()