+++ django_spire/core/management/commands/spire_startapp_pkg/template/app/tests/test_views/test_form_views.py
@@ -1,6 +1,0 @@
-from django.test import TestCase
-
-
-class SpireChildAppViewTestCase(TestCase):
-    def setUp(self):
-        super().setUp()