+++ django_spire/core/management/commands/spire_startapp_pkg/template/app/urls/__init__.py
@@ -1,9 +1,0 @@
-# from django.urls.conf import path, include
-#
-#
-# app_name = 'spirechildapp'
-#
-# urlpatterns = [
-#     path('page/', include('module.urls.page_urls', namespace='page')),
-#     path('form/', include('module.urls.form_urls', namespace='form')),
-# ]