+++ django_spire/knowledge/collection/models.py
@@ -1,9 +1,11 @@
 from django.db import models
 
+from django_spire.auth.group.models import AuthGroup
 from django_spire.contrib.ordering.mixins import OrderingModelMixin
 from django_spire.history.mixins import HistoryModelMixin
 from django_spire.knowledge.collection.querysets import CollectionQuerySet
-from django_spire.knowledge.collection.services.service import CollectionService
+from django_spire.knowledge.collection.services.service import CollectionGroupService, \
+    CollectionService
 
 
 class Collection(HistoryModelMixin, OrderingModelMixin):
@@ -29,3 +31,23 @@ class Collection(HistoryModelMixin, OrderingModelMixin):
         verbose_name = 'Collection'
         verbose_name_plural = 'Collections'
         db_table = 'django_spire_knowledge_collection'
+        permissions = [
+            ('can_access_all_collections', 'Can Access All Collections'),
+            ('can_change_collection_groups', 'Can Change Collection Groups')
+        ]
+
+
+class CollectionGroup(models.Model):
+    collection = models.ForeignKey(
+        Collection,
+        on_delete=models.CASCADE,
+        related_name='groups',
+        related_query_name='group',
+    )
+    auth_group = models.ForeignKey(
+        AuthGroup,
+        on_delete=models.CASCADE,
+        related_name='collection_groups',
+        related_query_name='collection_group',
+    )
+    services = CollectionGroupService()